Move DNS resolutions internal details in its own file (pacarser_dns.c)
